update feature added

+++ script.js
@@ -25,8 +25,16 @@ class UI{
         <td class="text-xl font-semibold p-4 ">${book.title}</td>
         <td class="text-xl font-semibold p-4 ">${book.author}</td>
         <td class="text-xl font-semibold p-4 ">${book.isbn}</td>
-        <td class="text-xl p-4 "><a href="#" class="bg-red-500 text-white p-2
-        delete">X</a></td>
+        <td class="text-xl p-4 ">
+            <a href="#" class="bg-red-500 text-white p-2 delete">X</a>
+        </td>
+        <td class="text-xl p-4 ">
+            <a href="#" class=" edit ">
+                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" class="size-12 rounded-lg border border-gray-300 p-2">
+                <path d="M21.731 2.269a2.625 2.625 0 0 0-3.712 0l-1.157 1.157 3.712 3.712 1.157-1.157a2.625 2.625 0 0 0 0-3.712ZM19.513 8.199l-3.712-3.712-12.15 12.15a5.25 5.25 0 0 0-1.32 2.214l-.8 2.685a.75.75 0 0 0 .933.933l2.685-.8a5.25 5.25 0 0 0 2.214-1.32L19.513 8.2Z" />
+                </svg>
+            </a>
+        </td>
         `;
 
         list.appendChild(row);
@@ -56,7 +64,35 @@ class UI{
         setTimeout(() => document.querySelector(".alert").remove(), 3000)
     };
 
-    
+    static editBook(el){
+            const row = el.closest("tr");
+
+            // Get the current values from the table row
+            const title = row.cells[0].innerText;
+            const author = row.cells[1].innerText;
+            const isbn = row.cells[2].innerText;
+
+            // Fill the inputs 
+            document.querySelector("#title").value = title;
+            document.querySelector("#author").value = author;
+            document.querySelector("#isbn").value = isbn;
+
+            document.querySelector("#isbn").readOnly = true;
+
+            const submitBtn = document.querySelector('input[type = "submit"]');
+            submitBtn.value = "Update Book";
+
+        
+    }
+
+
+    static resetForm(){
+        this.clearFields();
+        document.querySelector("#isbn").readOnly = false;
+        const submitBtn = document.querySelector('input[type = "submit"]');
+        submitBtn.value = "Add Book";
+
+    }
 }
 
 // Storage class : Handles storage
@@ -88,6 +124,17 @@ class Store{
 
         localStorage.setItem("books", JSON.stringify(books));
     }
+
+    static updateBook(updatedBook){
+        const books = Store.getBooks();
+        books.forEach((book, index) => {
+            if(book.isbn === updatedBook.isbn){
+                books[index] = updatedBook;
+            }
+        })
+
+        localStorage.setItem("books", JSON.stringify(books));
+    }
 }
 
 // Event : Display Books
@@ -102,6 +149,7 @@ document.querySelector("#book-form").addEventListener("submit", (e) => {
     const title = document.querySelector("#title").value;
     const author = document.querySelector("#author").value;
     const isbn = document.querySelector("#isbn").value;
+    const submitBtnValue = document.querySelector('input[type="submit"]').value;
 
     // Validate form
     if(title === "" || author === "" || isbn === ""){
@@ -109,30 +157,42 @@ document.querySelector("#book-form").addEventListener("submit", (e) => {
     }else{
         //Intatiate Book
         const book = new Book(title, author, isbn);
+        if(submitBtnValue === "Add Book"){
+             // Add Book to UI
+                UI.addBookToList(book);
 
-        // Add Book to UI
-        UI.addBookToList(book);
+                // Add Book to store
+                Store.addBook(book);
 
-        // Add Book to store
-        Store.addBook(book);
-
-        // Success message
-        UI.showAlert("Book Added", "green")
-
-        //Clear Fields
-        UI.clearFields();
+                // Success message
+                UI.showAlert("Book Added", "green")
+        }else{
+            // Edit Mode
+            Store.updateBook(book);
+            document.querySelector('#book-list').innerHTML = '';
+            UI.displayBooks();
+            UI.showAlert('Book Updated', 'green');
+            UI.resetForm();
+        }
+        
     }
-
-    
 })
 
 // Event : Remove a Book
 document.querySelector("#book-list").addEventListener("click", (e) => {
-    UI.deleteBook(e.target)
+    
+    if(e.target.classList.contains("delete")){
+        UI.deleteBook(e.target)
 
-    // Remove from store
-    Store.removeBook(e.target.parentElement.previousElementSibling.textContent);
+        // Remove from store
+        Store.removeBook(e.target.parentElement.previousElementSibling.textContent);
 
-    // Removed message
-    UI.showAlert("Book Removed", "green")
-})
+        // Removed message
+        UI.showAlert("Book Removed", "green")
+    }
+    //edit
+    const editBtn = e.target.closest(".edit");
+    if(editBtn){
+        UI.editBook(editBtn);
+    }
+});
